(start_server): fix printf format
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@8718 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ kadmin/kadm_conn.c
@@ -211,7 +211,7 @@ start_server(krb5_context context)
 	tmp = realloc(socks, (num_socks + i) * sizeof(*socks));
 	if(tmp == NULL) {
 	    krb5_warnx(context, "failed to reallocate %lu bytes", 
-		       (num_socks + i) * sizeof(*socks));
+		       (unsigned long)(num_socks + i) * sizeof(*socks));
 	    continue;
 	}
 	socks = tmp;
